comment

+++ app/Http/Controllers/RevisionPlanController.php
@@ -3466,10 +3466,10 @@ class RevisionPlanController extends Controller
                     'project_title' => $paps_title_desc,
                     'implementing_office' => $imp_office ? $imp_office : optional(optional($plan)->office)->FFUNCTION,
                     'expected_output' =>  $expected_outputs,
-                    'total_mooe' => number_format($total_mooe, 2, '.', ''),
-                    'total_ps' => number_format($total_ps, 2, '.', ''),
-                    'total_co' => number_format($total_co, 2, '.', ''),
-                    'total_fe' => number_format($total_fe, 2, '.', ''),
+                    'total_mooe' => (float) number_format($total_mooe, 2, '.', ''),
+                    'total_ps' => (float) number_format($total_ps, 2, '.', ''),
+                    'total_co' => (float) number_format($total_co, 2, '.', ''),
+                    'total_fe' => (float) number_format($total_fe, 2, '.', ''),
                     'ccet_code' => $ccetCode,
                     'ccet_code_mitigation' => $ccet_code_mitigation,
                     'ccet_code_adaptation' => $ccet_code_adaptation,
@@ -3746,68 +3746,68 @@ class RevisionPlanController extends Controller
                 $sf = $item['source_of_funds'];
                 $sector = $item['sector'];
                 if($sf=='gen_fund' && $sector=='General Public Services Sector'){
-                    $item['grand_total_mooe'] = number_format($general_public_services_total_mooe, 2, '.', '');
-                    $item['grand_total_ps']   = number_format($general_public_services_total_ps, 2, '.', '');
-                    $item['grand_total_co']   = number_format($general_public_services_total_co, 2, '.', '');
-                    $item['grand_total_fe']   = number_format($general_public_services_total_fe, 2, '.', '');
-                    $item['total_ccet_code_adaptation'] = number_format($general_public_services_ccet_code_adaptation, 2, '.', '');
-                    $item['total_ccet_code_mitigation'] = number_format($general_public_services_ccet_code_mitigation, 2, '.', '');
+                    $item['grand_total_mooe'] = (float) number_format($general_public_services_total_mooe, 2, '.', '');
+                    $item['grand_total_ps']   = (float) number_format($general_public_services_total_ps, 2, '.', '');
+                    $item['grand_total_co']   = (float) number_format($general_public_services_total_co, 2, '.', '');
+                    $item['grand_total_fe']   = (float) number_format($general_public_services_total_fe, 2, '.', '');
+                    $item['total_ccet_code_adaptation'] = (float) number_format($general_public_services_ccet_code_adaptation, 2, '.', '');
+                    $item['total_ccet_code_mitigation'] = (float) number_format($general_public_services_ccet_code_mitigation, 2, '.', '');
                 }
                 elseif($sf=='gen_fund' && $sector=='Economic Services'){
-                    $item['grand_total_mooe'] = number_format($economic_services_total_mooe, 2, '.', '');
-                    $item['grand_total_ps']   = number_format($economic_services_total_ps, 2, '.', '');
-                    $item['grand_total_co']   = number_format($economic_services_total_co, 2, '.', '');
-                    $item['grand_total_fe']   = number_format($economic_services_total_fe, 2, '.', '');
-                    $item['total_ccet_code_adaptation'] = number_format($economic_services_ccet_code_adaptation, 2, '.', '');
-                    $item['total_ccet_code_mitigation'] = number_format($economic_services_ccet_code_mitigation, 2, '.', '');
+                    $item['grand_total_mooe'] = (float) number_format($economic_services_total_mooe, 2, '.', '');
+                    $item['grand_total_ps']   = (float) number_format($economic_services_total_ps, 2, '.', '');
+                    $item['grand_total_co']   = (float) number_format($economic_services_total_co, 2, '.', '');
+                    $item['grand_total_fe']   = (float) number_format($economic_services_total_fe, 2, '.', '');
+                    $item['total_ccet_code_adaptation'] = (float) number_format($economic_services_ccet_code_adaptation, 2, '.', '');
+                    $item['total_ccet_code_mitigation'] = (float) number_format($economic_services_ccet_code_mitigation, 2, '.', '');
                 }
                 elseif($sf=='gen_fund' && $sector=='Social Services Sector'){
-                    $item['grand_total_mooe'] = number_format($social_services_total_mooe, 2, '.', '');
-                    $item['grand_total_ps']   = number_format($social_services_total_ps, 2, '.', '');
-                    $item['grand_total_co']   = number_format($social_services_total_co, 2, '.', '');
-                    $item['grand_total_fe']   = number_format($social_services_total_fe, 2, '.', '');
-                    $item['total_ccet_code_adaptation'] = number_format($social_services_ccet_code_adaptation, 2, '.', '');
-                    $item['total_ccet_code_mitigation'] = number_format($social_services_ccet_code_mitigation, 2, '.', '');
+                    $item['grand_total_mooe'] = (float) number_format($social_services_total_mooe, 2, '.', '');
+                    $item['grand_total_ps']   = (float) number_format($social_services_total_ps, 2, '.', '');
+                    $item['grand_total_co']   = (float) number_format($social_services_total_co, 2, '.', '');
+                    $item['grand_total_fe']   = (float) number_format($social_services_total_fe, 2, '.', '');
+                    $item['total_ccet_code_adaptation'] = (float) number_format($social_services_ccet_code_adaptation, 2, '.', '');
+                    $item['total_ccet_code_mitigation'] = (float) number_format($social_services_ccet_code_mitigation, 2, '.', '');
                 }
                 elseif($sf=='gen_fund' && $sector=='Other Services'){
-                    $item['grand_total_mooe'] = number_format($other_services_total_mooe, 2, '.', '');
-                    $item['grand_total_ps']   = number_format($other_services_total_ps, 2, '.', '');
-                    $item['grand_total_co']   = number_format($other_services_total_co, 2, '.', '');
-                    $item['grand_total_fe']   = number_format($other_services_total_fe, 2, '.', '');
-                    $item['total_ccet_code_adaptation'] = number_format($other_services_ccet_code_adaptation, 2, '.', '');
-                    $item['total_ccet_code_mitigation'] = number_format($other_services_ccet_code_mitigation, 2, '.', '');
+                    $item['grand_total_mooe'] = (float) number_format($other_services_total_mooe, 2, '.', '');
+                    $item['grand_total_ps']   = (float) number_format($other_services_total_ps, 2, '.', '');
+                    $item['grand_total_co']   = (float) number_format($other_services_total_co, 2, '.', '');
+                    $item['grand_total_fe']   = (float) number_format($other_services_total_fe, 2, '.', '');
+                    $item['total_ccet_code_adaptation'] = (float) number_format($other_services_ccet_code_adaptation, 2, '.', '');
+                    $item['total_ccet_code_mitigation'] = (float) number_format($other_services_ccet_code_mitigation, 2, '.', '');
                 }
                 elseif($sf=='dev'){
-                    $item['grand_total_mooe'] = number_format($dev_total_mooe, 2, '.', '');
-                    $item['grand_total_ps']   = number_format($dev_total_ps, 2, '.', '');
-                    $item['grand_total_co']   = number_format($dev_total_co, 2, '.', '');
-                    $item['grand_total_fe']   = number_format($dev_total_fe, 2, '.', '');
-                    $item['total_ccet_code_adaptation'] = number_format($dev_ccet_code_adaptation, 2, '.', '');
-                    $item['total_ccet_code_mitigation'] = number_format($dev_ccet_code_mitigation, 2, '.', '');
+                    $item['grand_total_mooe'] = (float) number_format($dev_total_mooe, 2, '.', '');
+                    $item['grand_total_ps']   = (float) number_format($dev_total_ps, 2, '.', '');
+                    $item['grand_total_co']   = (float) number_format($dev_total_co, 2, '.', '');
+                    $item['grand_total_fe']   = (float) number_format($dev_total_fe, 2, '.', '');
+                    $item['total_ccet_code_adaptation'] = (float) number_format($dev_ccet_code_adaptation, 2, '.', '');
+                    $item['total_ccet_code_mitigation'] = (float) number_format($dev_ccet_code_mitigation, 2, '.', '');
                 }
                 elseif($sf=='ldrrmf'){
-                    $item['grand_total_mooe'] = number_format($ldrrmf_total_mooe, 2, '.', '');
-                    $item['grand_total_ps']   = number_format($ldrrmf_total_ps, 2, '.', '');
-                    $item['grand_total_co']   = number_format($ldrrmf_total_co, 2, '.', '');
-                    $item['grand_total_fe']   = number_format($ldrrmf_total_fe, 2, '.', '');
-                    $item['total_ccet_code_adaptation'] = number_format($ldrrmf_ccet_code_adaptation, 2, '.', '');
-                    $item['total_ccet_code_mitigation'] = number_format($ldrrmf_ccet_code_mitigation, 2, '.', '');
+                    $item['grand_total_mooe'] = (float) number_format($ldrrmf_total_mooe, 2, '.', '');
+                    $item['grand_total_ps']   = (float) number_format($ldrrmf_total_ps, 2, '.', '');
+                    $item['grand_total_co']   = (float) number_format($ldrrmf_total_co, 2, '.', '');
+                    $item['grand_total_fe']   = (float) number_format($ldrrmf_total_fe, 2, '.', '');
+                    $item['total_ccet_code_adaptation'] = (float) number_format($ldrrmf_ccet_code_adaptation, 2, '.', '');
+                    $item['total_ccet_code_mitigation'] = (float) number_format($ldrrmf_ccet_code_mitigation, 2, '.', '');
                 }
                 elseif($sf=='other'){
-                    $item['grand_total_mooe'] = number_format($other_total_mooe, 2, '.', '');
-                    $item['grand_total_ps']   = number_format($other_total_ps, 2, '.', '');
-                    $item['grand_total_co']   = number_format($other_total_co, 2, '.', '');
-                    $item['grand_total_fe']   = number_format($other_total_fe, 2, '.', '');
-                    $item['total_ccet_code_adaptation'] = number_format($other_ccet_code_adaptation, 2, '.', '');
-                    $item['total_ccet_code_mitigation'] = number_format($other_ccet_code_mitigation, 2, '.', '');
+                    $item['grand_total_mooe'] = (float) number_format($other_total_mooe, 2, '.', '');
+                    $item['grand_total_ps']   = (float) number_format($other_total_ps, 2, '.', '');
+                    $item['grand_total_co']   = (float) number_format($other_total_co, 2, '.', '');
+                    $item['grand_total_fe']   = (float) number_format($other_total_fe, 2, '.', '');
+                    $item['total_ccet_code_adaptation'] = (float) number_format($other_ccet_code_adaptation, 2, '.', '');
+                    $item['total_ccet_code_mitigation'] = (float) number_format($other_ccet_code_mitigation, 2, '.', '');
                 }
                 else{
-                    $item['grand_total_mooe'] = number_format(0, 2, '.', '');
-                    $item['grand_total_ps']   = number_format(0, 2, '.', '');
-                    $item['grand_total_co']   = number_format(0, 2, '.', '');
-                    $item['grand_total_fe']   = number_format(0, 2, '.', '');
-                    $item['total_ccet_code_adaptation'] = number_format(0, 2, '.', '');
-                    $item['total_ccet_code_mitigation'] = number_format(0, 2, '.', '');
+                    $item['grand_total_mooe'] = (float) number_format(0, 2, '.', '');
+                    $item['grand_total_ps']   = (float) number_format(0, 2, '.', '');
+                    $item['grand_total_co']   = (float) number_format(0, 2, '.', '');
+                    $item['grand_total_fe']   = (float) number_format(0, 2, '.', '');
+                    $item['total_ccet_code_adaptation'] = (float) number_format(0, 2, '.', '');
+                    $item['total_ccet_code_mitigation'] = (float) number_format(0, 2, '.', '');
                 }
                 // $item['grand_total_mooe'] = number_format($general_public_services_total_mooe, 2, '.', '');
                 // $item['grand_total_ps']   = number_format($general_public_services_total_ps, 2, '.', '');
